done checkout page

+++ Pharmacy-admin/src/main/java/com/medibox/admin/reprository/OrderDetailsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.medibox.admin.model.OrderDetails;
 
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
 
 	
 	@Query("select u from OrderDetails u where u.orderdetailId=?1")
